Updated generate_wv_url
Size, in the WV url is the FOV of the image in arcseconds while the size in the wise_view_parameters is a pixel side length. So, using the known pixel ratio of the WISE data (~2.75 arc-seconds per pixel), I convert from the pixel side-length to the equivalent number of arc-seconds to generate the url.

Updated to version 0.6