Refactored wv.py into a WiseViewQuery Object
The functionality of wv.py is now implemented into a WiseViewQuery object.

This allows the JSON response to be stored and repeatedly accessed without the need of re-requesting from the AWS server. Since the get requests are the most time intensive part of the program, this was necessary to avoid re-requesting.