Small changes
Added a specific class methods for converting between FOVs and Pixel Sizes.

Created variables and comments for the parameters in the WiseView URL but are not parameters in the wise_view_parameters dictionary. This is just so its easier to modify them later on if necessary and visually see what they do.

Changed default values for some parameters:
window = 0.5 -> window = 1
speed = 125 -> speed = 150
gaia = 0 -> gaia = 1